fix: split function
