+++ packages/core/src/api/CheckBLEFirmwareRelease.ts
@@ -1,7 +1,7 @@
 import { BaseMethod } from './BaseMethod';
 
 import { UI_REQUEST } from '../constants/ui-request';
-import { getBleFirmwareReleaseInfo } from './firmware/releaseHelper';
+import { getBleFirmwareReleaseInfo } from './firmware/utils/releaseHelper';
 
 export default class CheckBLEFirmwareRelease extends BaseMethod {
   init() {
