+++ packages/core/src/api/firmware/uploadFirmware.ts
@@ -1,70 +1,15 @@
 import semver from 'semver';
-import { blake2s } from '@noble/hashes/blake2s';
 import JSZip from 'jszip';
 import { ERRORS, HardwareErrorCode } from '@onekeyfe/hd-shared';
-import { Success } from '@onekeyfe/hd-transport';
-import {
-  wait,
-  getDeviceBootloaderVersion,
-  getDeviceType,
-  LoggerNames,
-  getLogger,
-} from '../../utils';
-import { DEVICE, CoreMessage, createUiMessage, UI_REQUEST } from '../../events';
+import { wait, getDeviceBootloaderVersion, getDeviceType } from '../../utils';
+import { CoreMessage } from '../../events';
 import { PROTO } from '../../constants';
 import type { Device } from '../../device/Device';
-import type { TypedCall, TypedResponseMessage } from '../../device/DeviceCommands';
-import { DeviceModelToTypes, KnownDevice } from '../../types';
-import { bytesToHex } from '../helpers/hexUtils';
-import { DataManager } from '../../data-manager';
-import { DevicePool } from '../../device/DevicePool';
-import {
-  rebootDevice,
-  createFolder,
-  REBOOT_TYPE,
-  NEW_BOOT_UPRATE_FIRMWARE_VERSION,
-} from './bootloaderHelper';
-
-const SESSION_ERROR = 'session not found';
-const INIT_DATA_CHUNK_SIZE = 16 * 1024;
-const Log = getLogger(LoggerNames.Core);
-
-// UI Helper Functions
-const postConfirmationMessage = (device: Device) => {
-  // only if firmware is already installed. fresh device does not require button confirmation
-  if (device.features?.firmware_present) {
-    device.emit(DEVICE.BUTTON, device, { code: 'ButtonRequest_FirmwareUpdate' });
-  }
-};
-
-const postProgressMessage = (
-  device: Device,
-  progress: number,
-  postMessage: (message: CoreMessage) => void
-) => {
-  postMessage(
-    createUiMessage(UI_REQUEST.FIRMWARE_PROGRESS, {
-      device: device.toMessageObject() as KnownDevice,
-      progress,
-    })
-  );
-};
-
-const postProgressTip = (
-  device: Device,
-  message: string,
-  postMessage: (message: CoreMessage) => void
-) => {
-  postMessage(
-    createUiMessage(UI_REQUEST.FIRMWARE_TIP, {
-      device: device.toMessageObject() as KnownDevice,
-      data: {
-        message,
-      },
-    })
-  );
-};
-
+import type { TypedCall } from '../../device/DeviceCommands';
+import { DeviceModelToTypes } from '../../types';
+import { createFolder, emmcCommonUpdateProcess } from './utils/typedCallHelper';
+import { NEW_BOOT_UPRATE_FIRMWARE_VERSION } from './utils/const';
+import { postConfirmationMessage, postProgressTip, postProgressMessage } from './utils/uiHelper';
 // Basic Utility Functions
 export const waitBleInstall = async (updateType: string) => {
   if (updateType === 'ble') {
@@ -160,203 +105,6 @@ export const uploadFirmware = async (
   }
 
   throw ERRORS.TypedError(HardwareErrorCode.RuntimeError, 'uploadFirmware: unknown device model');
-};
-
-export const updateResource = async (typedCall: TypedCall, fileName: string, data: ArrayBuffer) => {
-  const chunk = new Uint8Array(data.slice(0, Math.min(INIT_DATA_CHUNK_SIZE, data.byteLength)));
-  const digest = blake2s(chunk);
-
-  const res = await typedCall('ResourceUpdate', ['ResourceRequest', 'Success'], {
-    file_name: fileName,
-    data_length: data.byteLength,
-    initial_data_chunk: bytesToHex(chunk),
-    hash: bytesToHex(digest),
-  });
-
-  return processResourceRequest(typedCall, res, data);
-};
-
-export const updateResources = async (
-  typedCall: TypedCall,
-  postMessage: (message: CoreMessage) => void,
-  device: Device,
-  source: ArrayBuffer
-) => {
-  postProgressTip(device, 'UpdateSysResource', postMessage);
-
-  const zipData = await JSZip.loadAsync(source);
-  const files = Object.entries(zipData.files);
-
-  let progress = 0;
-  const stepProgress = 100 / files.length;
-
-  for (const [fileName, file] of files) {
-    const name = fileName.split('/').pop();
-    if (!file.dir && fileName.indexOf('__MACOSX') === -1 && name) {
-      const data = await file.async('arraybuffer');
-      await updateResource(typedCall, name, data);
-    }
-
-    progress += stepProgress;
-    postProgressMessage(device, Math.floor(progress), postMessage);
-  }
-
-  postProgressMessage(device, 100, postMessage);
-  postProgressTip(device, 'UpdateSysResourceSuccess', postMessage);
-  return true;
-};
-
-export const updateBootloader = async (
-  typedCall: TypedCall,
-  postMessage: (message: CoreMessage) => void,
-  device: Device,
-  source: ArrayBuffer
-) => {
-  postProgressTip(device, 'UpdateBootloader', postMessage);
-  postProgressMessage(device, Math.floor(0), postMessage);
-  await updateResource(typedCall, 'bootloader.bin', source);
-  postProgressMessage(device, Math.floor(100), postMessage);
-  postProgressTip(device, 'UpdateBootloaderSuccess', postMessage);
-  return true;
-};
-
-// Complex Process Functions
-const processResourceRequest = async (
-  typedCall: TypedCall,
-  res: TypedResponseMessage<'ResourceRequest'> | TypedResponseMessage<'Success'>,
-  data: ArrayBuffer
-): Promise<Success> => {
-  if (res.type === 'Success') {
-    return res.message;
-  }
-
-  const { offset, data_length } = res.message;
-
-  if (offset === undefined) {
-    throw new Error('offset is undefined');
-  }
-
-  const payload = new Uint8Array(
-    data.slice(offset, Math.min(offset + data_length, data.byteLength))
-  );
-  const digest = blake2s(payload);
-
-  const resourceAckParams = {
-    data_chunk: bytesToHex(payload),
-    hash: bytesToHex(digest),
-  };
-
-  const response = await typedCall('ResourceAck', ['ResourceRequest', 'Success'], {
-    ...resourceAckParams,
-  });
-  return processResourceRequest(typedCall, response, data);
-};
-
-/**
- * 在bootloader模式下更新文件
- * @param payload 文件数据
- * @param filePath 文件路径
- * @param manulProgress 手动进度
- */
-const emmcCommonUpdateProcess = async (
-  device: Device,
-  {
-    payload,
-    filePath,
-    manulProgress,
-  }: PROTO.FirmwareUpload & { filePath: string; manulProgress?: number }
-) => {
-  const env = DataManager.getSettings('env');
-  const perPackageSize = DataManager.isBleConnect(env) ? 16 : 128;
-  const chunkSize = 1024 * perPackageSize;
-  const totalChunks = Math.ceil(payload.byteLength / chunkSize);
-  let offset = 0;
-  for (let i = 0; i < totalChunks; i++) {
-    const chunkStart = i * chunkSize;
-    const chunkEnd = Math.min(chunkStart + chunkSize, payload.byteLength);
-    const chunkLength = chunkEnd - chunkStart;
-    const chunk = payload.slice(chunkStart, chunkEnd);
-    const overwrite = i === 0;
-    const progress = Math.round(((i + 1) / totalChunks) * 100);
-    const writeRes = await emmcFileWriteWithRetry(
-      device,
-      filePath,
-      chunkLength,
-      offset,
-      chunk,
-      overwrite,
-      manulProgress ?? progress
-    );
-    // @ts-expect-error
-    offset += writeRes.message.processed_byte;
-    postProgressMessage(device, progress, postMessage);
-  }
-};
-
-const emmcFileWriteWithRetry = async (
-  device: Device,
-  filePath: string,
-  chunkLength: number,
-  offset: number,
-  chunk: ArrayBuffer,
-  overwrite: boolean,
-  progress: number | null
-) => {
-  const writeFunc = async () => {
-    const typedCall = device.getCommands().typedCall.bind(device.getCommands());
-    // @ts-expect-error
-    const writeRes = await typedCall('EmmcFileWrite', 'EmmcFile', {
-      file: {
-        path: filePath,
-        len: chunkLength,
-        offset,
-        data: chunk,
-      },
-      overwrite,
-      append: offset !== 0,
-      ui_percentage: progress,
-    });
-    if (writeRes.type !== 'EmmcFile') {
-      // @ts-expect-error
-      if (writeRes.type === 'CallMethodError') {
-        if (((writeRes as any).message.error ?? '').indexOf(SESSION_ERROR) > -1) {
-          throw ERRORS.TypedError(HardwareErrorCode.RuntimeError, SESSION_ERROR);
-        }
-      }
-      throw ERRORS.TypedError(HardwareErrorCode.RuntimeError, 'emmc file write chunk once error');
-    }
-    return writeRes;
-  };
-
-  let retryCount = 10;
-  while (retryCount > 0) {
-    try {
-      const result = await writeFunc();
-      return result;
-    } catch (error) {
-      Log.error(`emmcWrite error: `, error);
-      retryCount--;
-      if (retryCount === 0) {
-        throw ERRORS.TypedError(HardwareErrorCode.RuntimeError, 'emmc file write firmware error');
-      }
-      const env = DataManager.getSettings('env');
-      if (DataManager.isBleConnect(env)) {
-        await wait(3000);
-        await device.deviceConnector?.acquire(device.originalDescriptor.id, null, true);
-        await device.initialize();
-      } else if (error.message.indexOf(SESSION_ERROR) > -1) {
-        const deviceDiff = await device.deviceConnector?.enumerate();
-        const devicesDescriptor = deviceDiff?.descriptors ?? [];
-        const { deviceList } = await DevicePool.getDevices(devicesDescriptor, undefined);
-        if (deviceList.length === 1 && deviceList[0]?.features?.bootloader_mode) {
-          device.updateFromCache(deviceList[0]);
-          await device.acquire();
-          device.getCommands().mainId = device.mainId ?? '';
-        }
-      }
-      await wait(3000);
-    }
-  }
 };
 
 const newTouchUpdateFirmwareProcess = async (
@@ -463,18 +211,4 @@ export const updateResourcesInBootloaderMode = async (
     postProgressTip(device, 'UpdateSysResourceSuccess', postMessage);
     return true;
   }
-};
-
-// TODO: 后续再加入在bootloader中更新firmware, res, bootlaoder
-export const updateBootloaderInBootloaderMode = async (
-  typedCall: TypedCall,
-  device: Device,
-  source: ArrayBuffer
-) => {
-  await emmcCommonUpdateProcess(device, {
-    payload: source,
-    filePath: '0:boot/bootloader.bin',
-  });
-  await rebootDevice(typedCall, REBOOT_TYPE.REBOOT_BOOTLOADER);
-  return true;
 };
