+++ packages/core/src/api/firmware/utils/const.ts
@@ -1,0 +1,11 @@
+// Constants
+export const NEW_BOOT_UPRATE_FIRMWARE_VERSION = '2.4.5';
+
+// Reboot constants
+export const REBOOT_TYPE = {
+  REBOOT_NORMAL: 0,
+  REBOOT_BOARDLOADER: 1,
+  REBOOT_BOOTLOADER: 2,
+};
+
+export const INIT_DATA_CHUNK_SIZE = 16 * 1024;
