+++ packages/core/src/events/firmware.ts
@@ -1,5 +1,8 @@
 import { MessageFactoryFn } from './utils';
-import { getBleFirmwareReleaseInfo, getFirmwareReleaseInfo } from '../api/firmware/releaseHelper';
+import {
+  getBleFirmwareReleaseInfo,
+  getFirmwareReleaseInfo,
+} from '../api/firmware/utils/releaseHelper';
 import type { KnownDevice as Device } from '../types/device';
 
 export const FIRMWARE_EVENT = 'FIRMWARE_EVENT';
